# DTOs/DocenteCurso.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace DTOs
{
    public class DocenteCurso
    {
        public int Id { get; set; }
        public string Cargo { get; set; }
        public int IDCurso { get; set; }
        public int IDDocente { get; set; }
        
    }
}