# WebApplication1/obj/Debug/net8.0/EndpointInfo/ApiDocenteCurso.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ApiDocenteCurso",
    "version": "1.0"
  },
  "paths": {
    "/docentecursos/{id}": {
      "get": {
        "tags": [
          "ApiDocenteCurso"
        ],
        "operationId": "GetDocenteCurso",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocenteCurso"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "ApiDocenteCurso"
        ],
        "operationId": "UpdateDocenteCurso",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocenteCurso"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "delete": {
        "tags": [
          "ApiDocenteCurso"
        ],
        "operationId": "DeleteDocenteCurso",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/docentecursos": {
      "get": {
        "tags": [
          "ApiDocenteCurso"
        ],
        "operationId": "GetAllDocenteCursos",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocenteCurso"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ApiDocenteCurso"
        ],
        "operationId": "AddDocenteCurso",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocenteCurso"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocenteCurso"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "DocenteCurso": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "cargo": {
            "type": "string",
            "nullable": true
          },
          "idCurso": {
            "type": "integer",
            "format": "int32"
          },
          "idDocente": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      }
    }
  }
}